// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferRunLogAnalysisViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */
import type { TransferRunLogStageAnalysisViewDTO } from "./transferRunLogStageAnalysisViewDTO";

export interface TransferRunLogAnalysisViewDTO {
  /** 路由数量。 */
  routeCount?: number;
  /** 来源数量。 */
  sourceCount?: number;
  /** 阶段统计列表。 */
  stageAnalyses?: TransferRunLogStageAnalysisViewDTO[];
  /** 目标数量。 */
  targetCount?: number;
  /** 日志总数。 */
  totalCount?: number;
}