// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferRunLogStatusCountViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */

export interface TransferRunLogStatusCountViewDTO {
  /** 状态数量。 */
  count?: number;
  /** 运行状态。 */
  runStatus?: string;
  /** 状态名称。 */
  statusLabel?: string;
}